Track and report seldepth
seldepth shows the 'actual' depth reached during the search if the
search is extended to explore further in 'interesting' lines.

+++ engine/src/uci/responses.rs
@@ -34,8 +34,8 @@ pub(super) enum RegistrationStatus {
 
 #[derive(Debug, Default)]
 pub(crate) struct InfoFields {
-    pub(super) depth: Option<u32>,
-    pub(super) seldepth: Option<u32>,
+    pub(super) depth: Option<u8>,
+    pub(super) seldepth: Option<u8>,
     pub(super) time: Option<Duration>,
     pub(super) nodes: Option<u32>,
     pub(super) pv: Option<Vec<Move>>,
